Add files via upload
This is the final commit for the CPE301 Final Project for the Spring 2024 semester. This project was done together in person over the span of multiple days so there are not multiple commits to Github. Rather, we are uploading it all at once. This is the group consisting of Adam Lei, Sumiran Marsani, and Nikhil Anil.